refactor(store): flatten envelope/ into envelope.rs
Merge envelope/pending.rs and envelope/persisted.rs into a single
envelope.rs at the crate root. Write-path PendingEnvelope and read-path
PersistedEnvelope form one conceptual unit — the on-the-wire event
representation — and live together now.

Public paths nexus_store::envelope::{PendingEnvelope, PersistedEnvelope,
pending_envelope, WithVersion, WithEventType, WithPayload} unchanged.
