--- crates/nexus-store/src/envelope/mod.rs
@@ -1,5 +1,0 @@
-mod pending;
-mod persisted;
-
-pub use pending::{PendingEnvelope, WithEventType, WithPayload, WithVersion, pending_envelope};
-pub use persisted::PersistedEnvelope;
